Refactor modal imports and create new modals for sharing workflows and invite links
- Updated import paths for MoveToProjectModal and WorkflowDetailsModal to reflect new structure.
- Created CreateInviteLinkModal for generating invite links for workflows.
- Implemented MoveToProjectModal for moving workflows between projects.
- Added ShareWorkflowModal for sharing workflows with clients.
- Introduced WorkflowDetailsModal for displaying detailed information about workflows.
- Enhanced workflow data types to include diagram support.
- Updated package.json to include Mermaid for diagram rendering.
