+++ web/components/agent-chat/blocks/interrupt/interrupt-block.tsx
