+++ web/components/agent-chat/blocks/text/message-bubbles.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import InlineWidget from "@/components/agent-chat/widgets/inline-widget";
+import Block from "@/components/agent-chat/blocks/block";
 import Icon from "@/components/misc/icon";
 import MarkdownRender from "@/components/misc/markdown-render";
 import type {
@@ -180,7 +180,7 @@ export function AIResponseCard({
             </div>
           )}
           {widgets.map((w, wi) => (
-            <InlineWidget key={wi} data={w} />
+            <Block key={wi} data={w} />
           ))}
         </div>
       </div>
@@ -239,7 +239,7 @@ export function ResponseCard({
           {widgets.length > 0 && (
             <div className="px-3 pb-3">
               {widgets.map((w, wi) => (
-                <InlineWidget key={wi} data={w} />
+                <Block key={wi} data={w} />
               ))}
             </div>
           )}
