+++ web/components/agent-chat/blocks/todo-list/todo-list-block.tsx
@@ -5,7 +5,7 @@ import { useTranslations } from "@/lib/hooks/use-translations";
 import type { TodoListProps } from "@/components/agent-chat/types";
 import { Spinner } from "@heroui/react";
 
-export function TodoList({ todos }: TodoListProps) {
+export function TodoListBlock({ todos }: TodoListProps) {
   const { getTranslations: t } = useTranslations();
   const completed = todos.filter((t) => t.status === "completed").length;
   const progress = todos.length > 0 ? (completed / todos.length) * 100 : 0;
