+++ web/components/agent-chat/input/knowledge-files.tsx
