+++ web/components/agent-chat/input/project-picker.tsx
