+++ web/components/agent-chat/layouts/agent-chat-layouts.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import ChatInputBar from "@/components/agent-chat/widgets/input/chat-input-bar";
-import ChatMessageArea from "@/components/agent-chat/widgets/message/chat-message-area";
+import ChatInputBar from "@/components/agent-chat/input/chat-input-bar";
+import ChatMessageArea from "@/components/agent-chat/blocks/text/text-block";
 import Icon from "@/components/misc/icon";
 import { Button, Tooltip } from "@heroui/react";
 import { motion } from "framer-motion";
