+++ web/package.json
@@ -51,6 +51,7 @@
       "langchain": "^1.2.25",
       "lodash": "^4.17.23",
       "material-icons": "^1.13.14",
+      "mermaid": "^11.14.0",
       "next": "^16.1.6",
       "next-auth": "^4.24.13",
       "next-themes": "^0.4.6",
